DX16th/demoMVP_meetingmaterials/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
python-dateutil>=2.8.0
